# Meeting model scope fixes