--- a/modules/meeting/app/models/meeting.rb
+++ b/modules/meeting/app/models/meeting.rb
@@ -56,11 +56,10 @@ class Meeting < ApplicationRecord
   scope :templated, -> { where(template: true) }
   scope :not_templated, -> { where(template: false) }
 
-  scope :cancelled, -> { where(state: :cancelled) }
-  scope :not_cancelled, -> { where.not(id: cancelled) }
+  scope :not_cancelled, -> { where.not.cancelled }
 
   scope :not_recurring, -> { where(recurring_meeting_id: nil) }
-  scope :recurring, -> { where.not(id: not_recurring) }
+  scope :recurring, -> { where.not(recurring_meeting_id: nil) }
 
   scope :from_tomorrow, -> { where(start_time: Date.tomorrow.beginning_of_day..) }
   scope :from_today, -> { where(start_time: Time.zone.today.beginning_of_day..) }